+++ src/components/screen/auth/auth.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import AuthScreen from "./auth-screen.jsx";
+import AuthScreen from "./auth.jsx";
 
 
 const noop = () => {};
